Editing minor changes
removed unused imports
deleted unneccessary file
rearranged files to another package
