--- src/engine/src/rabbitescape/engine/behaviours/ClimbingTest.java
@@ -1,14 +1,0 @@
-package rabbitescape.engine.behaviours;
-
-import static org.junit.Assert.*;
-
-import org.junit.Test;
-
-public class ClimbingTest {
-
-	@Test
-	public void test() {
-		fail("Not yet implemented");
-	}
-
-}
