+++ src/engine/src/rabbitescape/engine/behaviours/bridging/BridgingInCornerLeft.java
@@ -1,4 +1,4 @@
-package rabbitescape.engine.behaviours;
+package rabbitescape.engine.behaviours.bridging;
 
 import static rabbitescape.engine.Block.Material.EARTH;
 import static rabbitescape.engine.Block.Shape.BRIDGE_UP_LEFT;
@@ -8,12 +8,10 @@ import rabbitescape.engine.Rabbit;
 import rabbitescape.engine.World;
 import rabbitescape.engine.ChangeDescription.State;
 
-public class BridgingLeft implements BridgingState {
+public class BridgingInCornerLeft implements BridgingState {
 	@Override
-	public boolean moveRabbit( World world, Rabbit rabbit, State state )
-	{
-        
-		rabbit.x--;
+	public boolean moveRabbit( World world, Rabbit rabbit, State state ) {
+		rabbit.onSlope = true;
         world.changes.addBlock(
             new Block(
                 rabbit.x,
@@ -23,7 +21,6 @@ public class BridgingLeft implements BridgingState {
                 0
             )
         );
-
         return true;
 	}
 }
