+++ src/engine/src/rabbitescape/engine/behaviours/bridging/BridgingInCornerUpRight.java
@@ -1,15 +1,14 @@
-package rabbitescape.engine.behaviours;
+package rabbitescape.engine.behaviours.bridging;
 
 import static rabbitescape.engine.Block.Material.EARTH;
-import static rabbitescape.engine.Block.Shape.BRIDGE_UP_LEFT;
-
+import static rabbitescape.engine.Block.Shape.BRIDGE_UP_RIGHT;
 
 import rabbitescape.engine.Block;
 import rabbitescape.engine.Rabbit;
 import rabbitescape.engine.World;
 import rabbitescape.engine.ChangeDescription.State;
 
-public class BridgingInCornerUpLeft implements BridgingState {
+public class BridgingInCornerUpRight implements BridgingState {
 	@Override
 	public boolean moveRabbit( World world, Rabbit rabbit, State state ) {
 		rabbit.onSlope = true;
@@ -18,7 +17,7 @@ public class BridgingInCornerUpLeft implements BridgingState {
                 rabbit.x,
                 rabbit.y,
                 EARTH,
-                BRIDGE_UP_LEFT,
+                BRIDGE_UP_RIGHT,
                 0
             )
         );
