+++ src/engine/src/rabbitescape/engine/behaviours/bridging/BridgingLeft.java
@@ -1,25 +1,25 @@
-package rabbitescape.engine.behaviours;
+package rabbitescape.engine.behaviours.bridging;
 
 import static rabbitescape.engine.Block.Material.EARTH;
-import static rabbitescape.engine.Block.Shape.BRIDGE_UP_RIGHT;
+import static rabbitescape.engine.Block.Shape.BRIDGE_UP_LEFT;
 
 import rabbitescape.engine.Block;
 import rabbitescape.engine.Rabbit;
 import rabbitescape.engine.World;
 import rabbitescape.engine.ChangeDescription.State;
 
-public class BridgingUpRight implements BridgingState {
+public class BridgingLeft implements BridgingState {
 	@Override
 	public boolean moveRabbit( World world, Rabbit rabbit, State state )
 	{
-		rabbit.x++;
-        rabbit.y--;
+        
+		rabbit.x--;
         world.changes.addBlock(
             new Block(
                 rabbit.x,
                 rabbit.y,
                 EARTH,
-                BRIDGE_UP_RIGHT,
+                BRIDGE_UP_LEFT,
                 0
             )
         );
