+++ src/engine/src/rabbitescape/engine/behaviours/bridging/BridgingRight.java
@@ -1,26 +1,28 @@
-package rabbitescape.engine.behaviours;
+package rabbitescape.engine.behaviours.bridging;
 
 import static rabbitescape.engine.Block.Material.EARTH;
-import static rabbitescape.engine.Block.Shape.BRIDGE_UP_LEFT;
+import static rabbitescape.engine.Block.Shape.BRIDGE_UP_RIGHT;
 
 import rabbitescape.engine.Block;
 import rabbitescape.engine.Rabbit;
 import rabbitescape.engine.World;
 import rabbitescape.engine.ChangeDescription.State;
 
-public class BridgingInCornerLeft implements BridgingState {
+public class BridgingRight implements BridgingState {
 	@Override
-	public boolean moveRabbit( World world, Rabbit rabbit, State state ) {
-		rabbit.onSlope = true;
+	public boolean moveRabbit( World world, Rabbit rabbit, State state )
+	{
+		rabbit.x++;
         world.changes.addBlock(
             new Block(
                 rabbit.x,
                 rabbit.y,
                 EARTH,
-                BRIDGE_UP_LEFT,
+                BRIDGE_UP_RIGHT,
                 0
             )
         );
+
         return true;
 	}
 }
