+++ src/engine/src/rabbitescape/engine/behaviours/bridging/BridgingUpRight.java
@@ -1,4 +1,4 @@
-package rabbitescape.engine.behaviours;
+package rabbitescape.engine.behaviours.bridging;
 
 import static rabbitescape.engine.Block.Material.EARTH;
 import static rabbitescape.engine.Block.Shape.BRIDGE_UP_RIGHT;
@@ -8,10 +8,12 @@ import rabbitescape.engine.Rabbit;
 import rabbitescape.engine.World;
 import rabbitescape.engine.ChangeDescription.State;
 
-public class BridgingInCornerUpRight implements BridgingState {
+public class BridgingUpRight implements BridgingState {
 	@Override
-	public boolean moveRabbit( World world, Rabbit rabbit, State state ) {
-		rabbit.onSlope = true;
+	public boolean moveRabbit( World world, Rabbit rabbit, State state )
+	{
+		rabbit.x++;
+        rabbit.y--;
         world.changes.addBlock(
             new Block(
                 rabbit.x,
@@ -21,6 +23,7 @@ public class BridgingInCornerUpRight implements BridgingState {
                 0
             )
         );
+
         return true;
 	}
 }
